Rename the main root package
