--- app/src/main/java/com/maximapps/weather/mainScreen/usecases/common/DetailedForecast.kt
@@ -1,18 +1,0 @@
-package com.maximapps.weather.mainScreen.usecases.common
-
-import androidx.annotation.DrawableRes
-import java.util.Date
-
-const val UNDEFINED = "undefined"
-
-data class DetailedForecast(
-    val date: Date,
-    val temperature: Int,
-    val temperatureMin: Int,
-    val temperatureMax: Int,
-    val wind: Int,
-    val feelsLike: Int,
-    val weatherCondition: String,
-    @DrawableRes val weatherIcon: Int,
-    val forecastList: List<WeatherForecast>
-)
