--- app/src/main/java/com/maximapps/weather/mainScreen/usecases/common/WeatherData.kt
@@ -1,6 +1,0 @@
-package com.maximapps.weather.mainScreen.usecases.common
-
-data class WeatherData(
-    val cityName: String = "",
-    val detailedForecast: List<DetailedForecast> = emptyList()
-)
