--- app/src/main/java/com/maximapps/weather/mainScreen/usecases/fetchforecastbycoordinates/Coordinates.kt
@@ -1,6 +1,0 @@
-package com.maximapps.weather.mainScreen.usecases.fetchforecastbycoordinates
-
-/**
- * Device geographic coordinates
- */
-data class Coordinates(val latitude: Double, val longitude: Double)
