--- app/src/main/java/com/maximapps/weather/mainScreen/usecases/fetchforecastbycoordinates/LocationRepository.kt
@@ -1,13 +1,0 @@
-package com.maximapps.weather.mainScreen.usecases.fetchforecastbycoordinates
-
-import kotlinx.coroutines.flow.Flow
-
-/**
- * Interface defining methods for getting device location.
- */
-interface LocationRepository {
-    /**
-     * Provides device geographic coordinates.
-     */
-    fun getLocation(): Flow<Coordinates>
-}
