+++ app/src/main/java/com/maximapps/weather/mainscreen/di/NetworkModule.kt
@@ -1,0 +1,18 @@
+package com.maximapps.weather.mainscreen.di
+
+import com.maximapps.weather.mainscreen.repositories.weather.network.WeatherService
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+class NetworkModule {
+    @Provides
+    @Singleton
+    fun providesOpenWeatherService(retrofit: Retrofit): WeatherService =
+        retrofit.create(WeatherService::class.java)
+}
