+++ app/src/main/java/com/maximapps/weather/mainscreen/usecases/common/WeatherData.kt
@@ -1,0 +1,6 @@
+package com.maximapps.weather.mainscreen.usecases.common
+
+data class WeatherData(
+    val cityName: String = "",
+    val detailedForecast: List<DetailedForecast> = emptyList()
+)
