+++ app/src/main/java/com/maximapps/weather/mainscreen/usecases/common/WeatherForecast.kt
@@ -1,0 +1,12 @@
+package com.maximapps.weather.mainscreen.usecases.common
+
+import androidx.annotation.DrawableRes
+import java.util.Date
+
+data class WeatherForecast(
+    val date: Date,
+    val temperature: Int,
+    val minTemperature: Int,
+    val maxTemperature: Int,
+    @DrawableRes val weatherIcon: Int
+)
