+++ app/src/main/java/com/maximapps/weather/mainscreen/usecases/fetchforecastbycoordinates/Coordinates.kt
@@ -1,0 +1,6 @@
+package com.maximapps.weather.mainscreen.usecases.fetchforecastbycoordinates
+
+/**
+ * Device geographic coordinates
+ */
+data class Coordinates(val latitude: Double, val longitude: Double)
