+++ app/src/main/java/com/maximapps/weather/mainscreen/usecases/fetchforecastbycoordinates/LocationRepository.kt
@@ -1,0 +1,13 @@
+package com.maximapps.weather.mainscreen.usecases.fetchforecastbycoordinates
+
+import kotlinx.coroutines.flow.Flow
+
+/**
+ * Interface defining methods for getting device location.
+ */
+interface LocationRepository {
+    /**
+     * Provides device geographic coordinates.
+     */
+    fun getLocation(): Flow<Coordinates>
+}
