+++ app/src/test/java/com/maximapps/weather/AllTestsSuite.kt
@@ -1,7 +1,7 @@
 package com.maximapps.weather
 
-import com.maximapps.weather.mainScreen.repositories.WeatherRepositoryTest
-import com.maximapps.weather.mainScreen.ui.MainViewModelTest
+import com.maximapps.weather.mainscreen.repositories.WeatherRepositoryTest
+import com.maximapps.weather.mainscreen.ui.MainViewModelTest
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
